fixup! Simplify opening tabs in TabManager
